backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pydantic[email]
email-validator
